Move the generate_config methods into a folder structure, fix support for Redis and start support of SNS
