+++ Gemfile
@@ -1,3 +1,13 @@
-source "https://rubygems.org"
-gem 'aws-sdk', '~> 2.3.10'
+# frozen_string_literal: true
+
+source "http://rubygems.org"
+
+git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
+
+ruby '~> 2.0'
+
+# For JSON interaction
 gem 'jason'
+
+# For AWS interaction
+gem 'aws-sdk', '~> 3'
